refactor cli to separate functions
