+++ capella-order/capella/__init__.py
@@ -1,1 +1,4 @@
 __version__ = "1.1.0"
+
+
+from capella.capella import *
